requirements.txt
pandas
scikit-learn
PyYAML
mlflow
pytest